build.rs: Link bundled proj statically
When building against the bundled proj 7.0 version, the resulting binary
still links against the dynamic library, and is prone to failing with
linker errors if the proj library with the correct version is not on the
library path.

By linking statically we ensure that the bundled proj library is
actually used by the resulting binary, and this is also more likely to
match the intent of the depending crate when the bundled_proj feature is
enabled.